# Batch RECTESTS into one process