--- a/run_rectests.sh
+++ b/run_rectests.sh
@@ -6,7 +6,7 @@ if [ -z "$1" ]; then
 fi
 
 export BUILD_DIR="$1"
-OPENOMF_BIN=$(find "$BUILD_DIR" -name openomf -type f -executable -print -quit)
+OPENOMF_BIN=$(find "$BUILD_DIR" "(" -name openomf -or -name openomf.exe ")" -type f -executable -print -quit)
 if [ -z "$OPENOMF_BIN" ]; then
     echo "Could not find openomf executable from $BUILD_DIR" >&2
     exit 1