--- a/src/engine.c
+++ b/src/engine.c
@@ -333,6 +333,8 @@ void engine_run(engine_init_flags *init_flags) {
             has_dynamic = dynamic_wait > dyntick_ms;
             if(has_dynamic) {
                 game_state_dynamic_tick(gs, false);
+                // check if we need to replace the game state
+                game_state_check_for_new(&gs);
                 dynamic_wait -= dyntick_ms;
                 if(gs->delay > 0) {
                     log_debug("applying delay %d", gs->delay);