--- a/src/game/game_state.h
+++ b/src/game/game_state.h
@@ -45,6 +45,7 @@ void game_state_init_demo(game_state *gs);
 int game_state_ms_per_dyntick(game_state *gs);
 ticktimer *game_state_get_ticktimer(game_state *gs);
 bool game_state_hars_are_alive(game_state *gs);
+void game_state_rec_finished(game_state *gs);
 
 object *game_state_find_object(game_state *gs, uint32_t object_id);
 int game_state_find_objects(game_state *gs, vector *out, bool (*predicate)(const object *obj, void *user_data),