--- a/src/game/game_state_type.h
+++ b/src/game/game_state_type.h
@@ -64,6 +64,7 @@ typedef struct game_state_t {
     uint32_t int_tick; // never adjusted, used in ping calculation
     unsigned int role;
     unsigned int speed;
+    unsigned int rec_playback_id;
     engine_init_flags *init_flags;
 
     match_settings match_settings;