… rectests/SHOULDFAIL.REC