All 60 rectests are now executed in a single process.

Benchmarks (msvc debug builds on windows with address sanitizer):
```
4m9.633s (master)
1m18.524s (master, but with --warp turned on)
0m25.373s (this branch, running all the RECs in series)
```

As an added bonus, running all 60 rectests in series is enough frames to quickly exhaust our renderer atlas:
<img width="802" height="632" alt="image" src="https://github.com/user-attachments/assets/d672b3fc-7aa7-4f2c-bde1-d274a55b2565" />