@Nopey 2025-09-25
timings in CI:
master: 3m 47s (6m 30s incl. build, menu tests, etc..)
this branch: 31s (5m 1s incl build, menu tests, etc..)