@Nopey 2025-09-25
warp runs every dynamic tick.
the timing between dynamic ticks and static ticks is not guaranteed (on any speed).

if warp skips any dynamic ticks, that's a bug.