@Vagabond 2025-09-25
I believe warp IS buggy, but I don't have i

> warp runs every dynamic tick. the timing between dynamic ticks and static ticks is not guaranteed (on any speed).
> 
> if warp skips any dynamic ticks, that's a bug.

I believe warp IS buggy, I just can't prove it.